requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymupdf==1.24.10
pdfplumber==0.10.3
pypdf==3.17.4
reportlab==4.0.8
pillow==10.1.0
openpyxl==3.1.2
python-docx==1.1.0
pytesseract==0.3.10
